% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_distributions_app.R
\name{run_distributions_app}
\alias{run_distributions_app}
\title{Launch the probability distributions app}
\usage{
run_distributions_app(...)
}
\arguments{
\item{...}{Optional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}. By default,
statsapps opens apps in the system default web browser using
\code{launch.browser = TRUE}. Use \code{launch.browser = FALSE} to disable this
behavior.}
}
\value{
\code{run_distributions_app()} opens the Shiny app locally.
}
\description{
Opens an interactive Shiny app for exploring common probability distributions
used in BIOL 300.
}
\details{
All data were simulated and don't come from a specific study.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_distributions_app()
\dontshow{\}) # examplesIf}
}
