% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_linear_reg_app.R
\name{run_linear_reg_app}
\alias{run_linear_reg_app}
\title{Launch the simple linear regression app}
\usage{
run_linear_reg_app(...)
}
\arguments{
\item{...}{Optional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}. Most users can
ignore this.}
}
\value{
\code{run_linear_reg_app()} opens the Shiny app locally.
}
\description{
Opens an interactive Shiny app for exploring how the slope and intercept of a
linear model affect fitted values and residuals.
}
\details{
All data were simulated and don't come from a specific study.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_linear_reg_app()
\dontshow{\}) # examplesIf}
}
