% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permutation_app.R
\name{run_permutation_app}
\alias{run_permutation_app}
\title{Launch the permutation test app}
\usage{
run_permutation_app(...)
}
\arguments{
\item{...}{Optional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}. By default,
statsapps opens apps in the system default web browser using
\code{launch.browser = TRUE}. Use \code{launch.browser = FALSE} to disable this
behavior.}
}
\value{
\code{run_permutation_app()} opens the Shiny app locally.
}
\description{
\code{run_permutation_app()} opens an interactive Shiny app that demonstrates how
repeated random reassignment of observations can be used to build a null
distribution for a two-sample permutation test.
}
\details{
The app uses data on time to mating in female sagebrush crickets from
Johnson et al. (1999), as presented in \emph{The Analysis of Biological Data} by
Whitlock and Schluter.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_permutation_app()
\dontshow{\}) # examplesIf}
}
\references{
Johnson, J. C., T. M. Ivy, and S. K. Sakaluk. 1999. Female remating
propensity contingent on sexual cannibalism in sagebrush crickets,
\emph{Cyphoderris strepitans}: a mechanism of cryptic female choice.
\emph{Behavioral Ecology} 10: 227-233.

Whitlock, M. C., and D. Schluter. 2020. \emph{The Analysis of Biological Data}.
3rd ed. W. H. Freeman and Company.
}
