% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sums_squares_app.R
\name{run_sums_squares_app}
\alias{run_sums_squares_app}
\title{Launch the sums of squares app}
\usage{
run_sums_squares_app(...)
}
\arguments{
\item{...}{Optional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}. By default,
statsapps opens apps in the system default web browser using
\code{launch.browser = TRUE}. Use \code{launch.browser = FALSE} to disable this
behavior.}
}
\value{
\code{run_sums_squares_app()} opens the Shiny app locally.
}
\description{
Opens an interactive Shiny app for exploring how total variation can be
partitioned into group-level and within-group components in one-way ANOVA.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
run_sums_squares_app()
\dontshow{\}) # examplesIf}
}
